Fix some incompatibilities with msvc, remove some dead simd_util bit permutation methods - Do some shenanigans to fallback to 32 bit built-in popcoun tif 64 bit one is not available - Fix simd_bits randomization test assuming SIZE_MAX is a 64 bit value

+++ src/main.cc
@@ -23,7 +23,7 @@ int main(int argc, const char **argv) {
     SampleFormat format =
         format_values[find_enum_argument("-format", SAMPLE_FORMAT_01, format_names.size(), format_names.data(), argc, argv)];
     bool interactive = find_bool_argument("-repl", argc, argv);
-    int samples = find_int_argument("-shots", 1, 0, 1 << 30, argc, argv);
+    int samples = (size_t) find_int_argument("-shots", 1, 0, 1 << 30, argc, argv);
     const char *out_path = find_argument("-out", argc, argv);
     FILE *out;
     if (out_path == nullptr) {
